+++ src/pages/mainPage/MainPageStyle.jsx
